event for japanese capitulation
Added an event for an historical japanese capitulation
It would need later to be more dynamical, but now it's mainly a way to end the war between allies and Japan
Koreas should be handled later : now it's just under soviet/us occupation. A tag for North Korea would be mandatory.

Concerning the trigger, I didn't consider the nuke as relevant for historicaly backed considerations. I consider the soviet invasion far more relevant, so Soviet have to occupy Mandchouria + North Korea to trigger the event.