Reafirm eastern claims now gives lost german cores from WWI
Memel, Danzig, Pomerelia and Kattowice are german cores lost in the WWI. Like Alsace-Moselle, they should be treated as german core.
I didn't add them directly to their respectives provinces because of the games mechanics (bad relationship etc).
Poznan is considered as a claim and not a core because even in the days of Prussia/ German Empire, the region was never assimilated and german speaking people remain a slight minority there.